+++ src/functions/dependsOn.ts
@@ -13,8 +13,59 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-import type { DependencyInfoCollectionArg, DependencyInfoResolver, IClassDependencyItem, IDependencyInformation, IFunctionDependencyItem } from "../types";
+import type { DependencyInfoCollectionArg, DependencyInfoResolver, IClassDependencyItem, IDependencyInformation, IFunctionDependencyItem, IModuleDependencyItem, IStackInfo } from "../types";
 import type { Constructor, CreateDependsOnHelpersFunc, Func, Nilable } from "../types/internal";
+
+/**
+ * Adds dependency information for current app.
+ *
+ * @example
+ * ```
+ * import {
+ *   appDependsOn,
+ *   getDependencies
+ * } from "@egomobile/documentation"
+ *
+ * appDependsOn(
+ *   {
+ *     // your information here...
+ *   }
+ * )
+ *
+ * console.log(
+ *   getDependencies()
+ * )
+ * ```
+ *
+ * @param {IDependencyInformation|DependencyInfoResolver} infoOrResolver The dependency information or the function that resolves it.
+ * @param {Nilable<DependenciesArg>} [dependenciesOrResolver] The custom collection for the dependency info items or the function that resolves it.
+ * @param {Nilable<IStackInfo|false>} [stackInfo] Custom stack information.
+ *
+ * @throws ReferenceError No CommonJS based main module found.
+ */
+export function appDependsOn(
+    infoOrResolver: IDependencyInformation | DependencyInfoResolver,
+    dependenciesOrResolver?: Nilable<DependencyInfoCollectionArg>,
+    stackInfo?: Nilable<IStackInfo | false>
+) {
+    const mainModule = require.main;
+    if (!mainModule) {
+        throw new ReferenceError("no main module found");
+    }
+
+    const tryGetStackInfo = require("../utils/internal").tryGetStackInfo;
+
+    if (!stackInfo) {
+        stackInfo = tryGetStackInfo();
+    }
+
+    moduleDependsOn(
+        mainModule,
+        infoOrResolver,
+        dependenciesOrResolver,
+        stackInfo
+    );
+}
 
 /**
  * Adds dependency information for a class.
@@ -31,7 +82,7 @@ import type { Constructor, CreateDependsOnHelpersFunc, Func, Nilable } from "../
  * }
  *
  * classDependsOn(
- *   MyClass
+ *   MyClass,
  *   {
  *     // your information here...
  *   }
@@ -45,14 +96,21 @@ import type { Constructor, CreateDependsOnHelpersFunc, Func, Nilable } from "../
  * @param {Constructor<any>} classConstructor The class constructor.
  * @param {IDependencyInformation|DependencyInfoResolver} infoOrResolver The dependency information or the function that resolves it.
  * @param {Nilable<DependenciesArg>} [dependenciesOrResolver] The custom collection for the dependency info items or the function that resolves it.
+ * @param {Nilable<IStackInfo|false>} [stackInfo] Custom stack information.
  */
 export function classDependsOn(
     classConstructor: Constructor<any>,
     infoOrResolver: IDependencyInformation | DependencyInfoResolver,
-    dependenciesOrResolver?: Nilable<DependencyInfoCollectionArg>
+    dependenciesOrResolver?: Nilable<DependencyInfoCollectionArg>,
+    stackInfo?: Nilable<IStackInfo | false>
 ) {
     const createDependsOnHelpers: CreateDependsOnHelpersFunc =
         require("../utils/internal").createDependsOnHelpers;
+    const tryGetStackInfo = require("../utils/internal").tryGetStackInfo;
+
+    if (!stackInfo) {
+        stackInfo = tryGetStackInfo();
+    }
 
     const {
         addItem
@@ -63,7 +121,10 @@ export function classDependsOn(
         "type": "class"
     };
 
-    addItem(newClassItem);
+    addItem({
+        "existsIn": stackInfo ?? null,
+        "item": newClassItem
+    });
 }
 
 /**
@@ -81,7 +142,7 @@ export function classDependsOn(
  * }
  *
  * functionDependsOn(
- *   myFunction
+ *   myFunction,
  *   {
  *     // your information here...
  *   }
@@ -95,14 +156,21 @@ export function classDependsOn(
  * @param {Func} func The function.
  * @param {IDependencyInformation|DependencyInfoResolver} infoOrResolver The dependency information or the function that resolves it.
  * @param {Nilable<DependenciesArg>} [dependenciesOrResolver] The custom collection for the dependency info items or the function that resolves it.
+ * @param {Nilable<IStackInfo|false>} [stackInfo] Custom stack information.
  */
 export function functionDependsOn(
     func: Func,
     infoOrResolver: IDependencyInformation | DependencyInfoResolver,
-    dependenciesOrResolver?: Nilable<DependencyInfoCollectionArg>
+    dependenciesOrResolver?: Nilable<DependencyInfoCollectionArg>,
+    stackInfo?: Nilable<IStackInfo | false>
 ) {
     const createDependsOnHelpers: CreateDependsOnHelpersFunc =
         require("../utils/internal").createDependsOnHelpers;
+    const tryGetStackInfo = require("../utils/internal").tryGetStackInfo;
+
+    if (!stackInfo) {
+        stackInfo = tryGetStackInfo();
+    }
 
     const {
         addItem
@@ -113,5 +181,64 @@ export function functionDependsOn(
         "type": "function"
     };
 
-    addItem(newFunctionItem);
+    addItem({
+        "existsIn": stackInfo ?? null,
+        "item": newFunctionItem
+    });
+}
+
+/**
+ * Adds dependency information for a module.
+ *
+ * @example
+ * ```
+ * import {
+ *   moduleDependsOn,
+ *   getDependencies
+ * } from "@egomobile/documentation"
+ *
+ * moduleDependsOn(
+ *   require.main,
+ *   {
+ *     // your information here...
+ *   }
+ * )
+ *
+ * console.log(
+ *   getDependencies()
+ * )
+ * ```
+ *
+ * @param {any} module The module.
+ * @param {IDependencyInformation|DependencyInfoResolver} infoOrResolver The dependency information or the function that resolves it.
+ * @param {Nilable<DependenciesArg>} [dependenciesOrResolver] The custom collection for the dependency info items or the function that resolves it.
+ * @param {Nilable<IStackInfo|false>} [stackInfo] Custom stack information.
+ */
+export function moduleDependsOn(
+    module: any,
+    infoOrResolver: IDependencyInformation | DependencyInfoResolver,
+    dependenciesOrResolver?: Nilable<DependencyInfoCollectionArg>,
+    stackInfo?: Nilable<IStackInfo | false>
+) {
+    const createDependsOnHelpers: CreateDependsOnHelpersFunc =
+        require("../utils/internal").createDependsOnHelpers;
+    const tryGetStackInfo = require("../utils/internal").tryGetStackInfo;
+
+    if (!stackInfo) {
+        stackInfo = tryGetStackInfo();
+    }
+
+    const {
+        addItem
+    } = createDependsOnHelpers(infoOrResolver, dependenciesOrResolver);
+
+    const newModuleItem: IModuleDependencyItem = {
+        module,
+        "type": "module"
+    };
+
+    addItem({
+        "existsIn": stackInfo ?? null,
+        "item": newModuleItem
+    });
 }